move source code into src directory
